fixup, fixed bug in nearest neightbour

+++ src/resize_image.cpp
@@ -15,6 +15,8 @@ float Image::pixel_nearest(float x, float y, int c) const
   // TODO: Your code here
   int rx = round(x);
   int ry = round(y);
+  //printf("%f,%f\t%d,%d\n", x, y, rx, ry);
+
   return this->clamped_pixel(rx, ry, c);
   }
 
@@ -29,11 +31,33 @@ float Image::pixel_bilinear(float x, float y, int c) const
   
   
   // TODO: Your code here
-  
   NOT_IMPLEMENTED();
-  
-  
   return 0;
+
+ 
+
+//  int fx = floor(x);
+//  int cx = ceil(x);
+//  int fy = floor(y);
+//  int cy = ceil(y);
+//
+//
+//
+//  float vffx = (x - fx) * this->clamped_pixel(fx, fy, c);
+//  float vfcx = (cx - x) * this->clamped_pixel(cx, fy, c);
+//  float vfx  = vffx + vfcx;
+//
+//  float vcfx = (x - fx) * this->clamped_pixel(fx, cy, c);
+//  float vccx = (cx - x) * this->clamped_pixel(cx, cy, c);
+//  float vcx = vcfx + vccx;
+//
+//
+//  float vfy = (y - fy) * vfx;
+//  float vcy = (cy - y) * vcx;
+//
+//  float v = vfy + vcy;
+//
+//  return v;
   }
 
 // HW1 #1
@@ -43,8 +67,9 @@ float Image::pixel_bilinear(float x, float y, int c) const
 Image nearest_resize(const Image& im, int w, int h)
   {
   Image ret(w,h,im.c);
-  float scale_x = w / im.w;
-  float scale_y = h / im.h;
+
+  float scale_x = (float) w / im.w;
+  float scale_y = (float) h / im.h;
 
   for (int c = 0; c < ret.c; c ++) {
       for (int y = 0; y < h; y ++) {
